pyproject.toml
[tool.poetry]
name = "viiteri"
version = "0.1.0"
description = "Ryhmän JEEOVI ohtu miniprojekti"
license="MIT"
authors = ["3nd3r1 <ender.spam.guru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
invoke = "^2.2.0"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
rpaframework = "^27.7.0"
bibtexparser = "^1.4.1"
python-doi = "^0.2.0"


[tool.poetry.dev-dependencies]
pylint = "^3.0.2"
robotframework = "^6.1.1"
coverage = "^7.3.2"
pytest = "^7.4.3"
robotframework-seleniumlibrary = "^6.1.3"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"